Add files via upload
Course 2 End-of-course project: Inspect and analyze data
In this activity, you will examine data provided and prepare it for analysis. This activity will help ensure the information is,

Ready to answer questions and yield insights

Ready for visualizations

Ready for future hypothesis testing and statistical methods
The purpose of this project is to investigate and understand the data provided.

The goal is to use a dataframe contructed within Python, perform a cursory inspection of the provided dataset, and inform team members of your findings.

This activity has three parts:

Part 1: Understand the situation

Prepare to understand and organize the provided taxi cab dataset and information.
Part 2: Understand the data

Create a pandas dataframe for data learning, future exploratory data analysis (EDA), and statistical activities.

Compile summary information about the data to inform next steps.

Part 3: Understand the variables

Use insights from your examination of the summary data to guide deeper investigation into specific variables.

Follow the instructions and answer the following questions to complete the activity. Then, you will complete an Executive Summary using the questions listed on the PACE Strategy Document.

Be sure to complete this activity before moving on. The next course item will provide you with a completed exemplar to compare to your own work.

Identify data types and relevant variables using Python